skycoin-lite: make the wasm cipher registrable from any wasm program
The registration lived in a main, so the only way to get the cipher into a
browser was to serve the binary built from src/skycoin-lite/wasm. Nothing about
it needed to be a program: it publishes two objects on the JavaScript global and
returns.

It moves to src/skycoin-lite/wasmcipher, exported as Register. The main is now a
call to it plus the block that keeps the program alive — Register itself does not
block, because a host carrying the cipher alongside other things has its own
reason to stay running.

What this is for: the wallet front end reaches the cipher only through
window.SkycoinCipher (cipher.provider.ts), so any wasm that calls Register is a
drop-in for skycoin-lite.wasm and the page cannot tell the difference. A visor
serving the wallet could carry it and, by doing the node requests itself, remove
the reason cmd/skycoin-web proxies /api at all. This commit does not do that; it
removes the packaging reason it could not be done.

Returning errors rather than panicking, in #3025, was the precondition. A
library that reports failure by panicking is not safe to link into a host doing
other work — under TinyGo a mistyped address would have trapped the host, not
just the cipher.

One behavioural note, recorded on buildVersion: the version reported through
SkycoinCipher.version describes the program that called Register, so a visor
carrying the cipher reports the visor's build. That is the useful answer, since
it identifies the module the browser actually loaded, but it will not always be
a skycoin commit.
