+++ src/skycoin-lite/wasm/main_wasm.go
@@ -3,179 +3,23 @@
 // Package main provides the wasm build of the browser cipher: the entry points
 // src/skycoin-web calls to derive addresses and sign transactions, plus the
 // verification helpers its cipher spec checks them with.
+//
+// The registration itself lives in src/skycoin-lite/wasmcipher, so a wasm
+// program that does more than this one — a visor serving the wallet, say — can
+// carry the same cipher by calling Register. This build is the cipher on its
+// own, which is what src/skycoin-web serves today.
 package main
 
 import (
-	"fmt"
-	"runtime/debug"
-	"syscall/js"
-
-	"github.com/skycoin/skycoin/src/skycoin-lite/liteclient"
+	"github.com/skycoin/skycoin/src/skycoin-lite/wasmcipher"
 )
 
-// buildVersion reports what this wasm was built from.
-//
-// The toolchain already records it: a `go build` inside a git work tree stamps
-// the module version and the vcs.* settings into the binary, so nothing needs
-// injecting through -ldflags. Reading it back is the only way to tell, from a
-// running wallet, which cipher it is actually holding — the wasm is a committed
-// artifact, so the commit it was built at is necessarily an earlier one than the
-// commit that carries it.
-//
-// modified is the interesting field. It means the working tree had uncommitted
-// changes when this was compiled, so no commit describes what is in it.
-//
-// TinyGo does not record any of this, and reports empty strings.
-func buildVersion() map[string]interface{} {
-	version := map[string]interface{}{
-		"version":  "",
-		"commit":   "",
-		"date":     "",
-		"modified": false,
-	}
-
-	info, ok := debug.ReadBuildInfo()
-	if !ok {
-		return version
-	}
-
-	version["version"] = info.Main.Version
-	for _, setting := range info.Settings {
-		switch setting.Key {
-		case "vcs.revision":
-			version["commit"] = setting.Value
-		case "vcs.time":
-			version["date"] = setting.Value
-		case "vcs.modified":
-			version["modified"] = setting.Value == "true"
-		}
-	}
-
-	return version
-}
-
-// errorResult is what every entry point returns when it could not do its job.
-// Callers check for the "error" key, so the shape has to be a map even when the
-// success value is a plain string.
-func errorResult(err interface{}) map[string]interface{} {
-	return map[string]interface{}{"error": fmt.Sprint(err)}
-}
-
-// guard runs fn and turns a panic into an {"error": ...} result.
-//
-// liteclient returns errors now, so nothing below is expected to panic and this
-// should never fire. It stays as a backstop for the cipher package underneath,
-// which still reports some failures that way.
-//
-// It cannot be relied on. Under TinyGo the recover does not fire at all and a
-// panic traps the whole module, leaving the cipher dead for the rest of the
-// page — which is why the errors are returned rather than recovered.
-func guard(fn func() (interface{}, error)) (result interface{}) {
-	defer func() {
-		if r := recover(); r != nil {
-			result = errorResult(r)
-		}
-	}()
-
-	value, err := fn()
-	if err != nil {
-		return errorResult(err)
-	}
-
-	return value
-}
-
-// arity wraps a function of n string arguments, rejecting short calls before
-// they reach the cipher.
-func arity(n int, name string, fn func(args []js.Value) (interface{}, error)) js.Func {
-	return js.FuncOf(func(_ js.Value, args []js.Value) interface{} {
-		if len(args) < n {
-			return errorResult(fmt.Sprintf("%s requires %d argument(s)", name, n))
-		}
-
-		return guard(func() (interface{}, error) { return fn(args) })
-	})
-}
-
 func main() {
-	// The wallet's own entry points.
-	skycoinCipher := js.Global().Get("Object").New()
+	wasmcipher.Register()
 
-	skycoinCipher.Set("generateAddress", arity(1, "generateAddress", func(args []js.Value) (interface{}, error) {
-		address, err := liteclient.GenerateAddress(args[0].String())
-		if err != nil {
-			return nil, err
-		}
-
-		return map[string]interface{}{
-			"nextSeed": address.NextSeed,
-			"secret":   address.Secret,
-			"public":   address.Public,
-			"address":  address.Address,
-		}, nil
-	}))
-
-	skycoinCipher.Set("prepareTransaction", arity(2, "prepareTransaction", func(args []js.Value) (interface{}, error) {
-		return liteclient.PrepareTransaction(args[0].String(), args[1].String())
-	}))
-
-	// prepareTransactionWithSignatures is used for hardware wallet signing.
-	skycoinCipher.Set("prepareTransactionWithSignatures", arity(3, "prepareTransactionWithSignatures", func(args []js.Value) (interface{}, error) {
-		return liteclient.PrepareTransactionWithSignatures(args[0].String(), args[1].String(), args[2].String())
-	}))
-
-	skycoinCipher.Set("version", buildVersion())
-
-	js.Global().Set("SkycoinCipher", skycoinCipher)
-
-	// The verification helpers. These exist so the browser cipher can be checked
-	// against src/cipher/testsuite/testdata — the same golden vectors the Go
-	// implementation is checked against — by
-	// src/skycoin-web/src/app/services/cipher.provider.lib.spec.ts. The GopherJS
-	// build publishes the same set as CipherExtras.
-	//
-	// Each verify* returns null when the check passes and an error string when it
-	// does not, rather than the int the underlying secp256k1 helpers use.
-	skycoinCipherExtras := js.Global().Get("Object").New()
-
-	skycoinCipherExtras.Set("verifyPubKeySignedHash", arity(3, "verifyPubKeySignedHash", func(args []js.Value) (interface{}, error) {
-		return nil, liteclient.VerifyPubKeySignedHash(args[0].String(), args[1].String(), args[2].String())
-	}))
-
-	skycoinCipherExtras.Set("verifyAddressSignedHash", arity(3, "verifyAddressSignedHash", func(args []js.Value) (interface{}, error) {
-		return nil, liteclient.VerifyAddressSignedHash(args[0].String(), args[1].String(), args[2].String())
-	}))
-
-	skycoinCipherExtras.Set("verifySignatureRecoverPubKey", arity(2, "verifySignatureRecoverPubKey", func(args []js.Value) (interface{}, error) {
-		return nil, liteclient.VerifySignatureRecoverPubKey(args[0].String(), args[1].String())
-	}))
-
-	skycoinCipherExtras.Set("verifySeckey", arity(1, "verifySeckey", func(args []js.Value) (interface{}, error) {
-		return nil, liteclient.VerifySeckey(args[0].String())
-	}))
-
-	skycoinCipherExtras.Set("verifyPubkey", arity(1, "verifyPubkey", func(args []js.Value) (interface{}, error) {
-		return nil, liteclient.VerifyPubkey(args[0].String())
-	}))
-
-	skycoinCipherExtras.Set("addressFromPubKey", arity(1, "addressFromPubKey", func(args []js.Value) (interface{}, error) {
-		return liteclient.AddressFromPubKey(args[0].String())
-	}))
-
-	skycoinCipherExtras.Set("addressFromSecKey", arity(1, "addressFromSecKey", func(args []js.Value) (interface{}, error) {
-		return liteclient.AddressFromSecKey(args[0].String())
-	}))
-
-	skycoinCipherExtras.Set("pubKeyFromSig", arity(2, "pubKeyFromSig", func(args []js.Value) (interface{}, error) {
-		return liteclient.PubKeyFromSig(args[0].String(), args[1].String())
-	}))
-
-	skycoinCipherExtras.Set("signHash", arity(2, "signHash", func(args []js.Value) (interface{}, error) {
-		return liteclient.SignHash(args[0].String(), args[1].String())
-	}))
-
-	js.Global().Set("SkycoinCipherExtras", skycoinCipherExtras)
-
-	// Keep the Go program running
+	// Keep the Go program running: JavaScript calls into the exported functions
+	// long after main would otherwise return, and exiting would take them with
+	// it. A host with its own reason to stay alive does not need this, which is
+	// why Register does not block.
 	<-make(chan struct{})
 }
